requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1 # Para cargar variables de entorno (SRT_URL, etc.)
pydantic==2.7.4 # FastAPI lo usa internamente